organize repo & directory tree
